requirements.txt
numpy
pillow
tensorflow
sklearn
matplotlib
